venuex-core/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.1
openai>=1.3.0
temporalio>=1.4.0
firecrawl-py>=0.0.8
firebase-admin>=6.2.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
httpx>=0.24.0